# perf(runtime): flatten arguments for write_file ops