~20% improvement in `await Deno.writeFile`

```
# This patch
time 151 ms rate 66225
time 157 ms rate 63694
time 152 ms rate 65789
time 151 ms rate 66225
time 152 ms rate 65789
time 152 ms rate 65789
```

```
# main
time 188 ms rate 53191
time 173 ms rate 57803
time 170 ms rate 58823
time 173 ms rate 57803
time 173 ms rate 57803
time 173 ms rate 57803
```